Google - Hard - Divide and Conquer
The core method for this problem is to divide those two lists into left parts and right parts, with i for nums1 and j for nums2, where i+j == (m+n)/2 or (m+n+1)/2. So, if we get i, we can get the corresponding value for j.

Then, we apply binary search for i.

The i, which is the number we want to find, is initialized as (l+r)/2. We need to find when we should update l and when we should update r.
1) when nums2[j-1] > nums[i], it menas that nums[i] is not big enough, so that we increase l with l = i+1
2) when nums1[i-1] > nums[j], it means that nums[i] is too large, so that we decrease r with r = i-1

The termination condition are nums1[i-1] < nums2[j] and nums2[j-1] < nums1[i]. Be careful about the edge situations: i=0, j=0, i=m, j=n, where m = len(nums1), n = len(nums2) and m < n.